Fix guessing blank table after first guess
Table validity is checked only when an item is placed on the crafting bench. This is fine for the first guess, because the table validity is false by default. However, after guessing validly and creating a new guessing table, the table validity isn't reset to false. This means the player can keep "guessing" (by clicking the crafting-output box) as long as they don't trigger the check by adding any items to the table.

Fixed by adding a validity check whenever the crafting output box is clicked before processing the guess.